tests: Move test_swf_approx to util/runner
